@@ -13,7 +13,7 @@
     "etc/prepare.js"
   ],
   "types": "bson.d.ts",
-  "version": "5.1.0",
+  "version": "5.3.0",
   "author": {
     "name": "The MongoDB NodeJS Team",
     "email": "dbx-node@mongodb.com"
@@ -26,37 +26,37 @@
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.33.7",
-    "@rollup/plugin-node-resolve": "^15.0.1",
-    "@rollup/plugin-typescript": "^10.0.1",
-    "@types/chai": "^4.3.4",
+    "@microsoft/api-extractor": "^7.34.7",
+    "@rollup/plugin-node-resolve": "^15.0.2",
+    "@rollup/plugin-typescript": "^11.1.0",
+    "@types/chai": "^4.3.5",
     "@types/mocha": "^10.0.1",
-    "@types/node": "^18.11.12",
-    "@types/sinon": "^10.0.13",
+    "@types/node": "^18.16.3",
+    "@types/sinon": "^10.0.14",
     "@types/sinon-chai": "^3.2.9",
-    "@typescript-eslint/eslint-plugin": "^5.46.0",
-    "@typescript-eslint/parser": "^5.46.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.2",
+    "@typescript-eslint/parser": "^5.59.2",
     "benchmark": "^2.1.4",
     "chai": "^4.3.7",
-    "chalk": "^5.1.2",
-    "eslint": "^8.29.0",
-    "eslint-config-prettier": "^8.5.0",
+    "chalk": "^5.2.0",
+    "eslint": "^8.39.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-no-bigint-usage": "file:./etc/eslint/no-bigint-usage",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-tsdoc": "^0.2.17",
-    "magic-string": "^0.27.0",
-    "mocha": "10.1.0",
-    "node-fetch": "^3.2.10",
+    "magic-string": "^0.30.0",
+    "mocha": "10.2.0",
+    "node-fetch": "^3.3.1",
     "nyc": "^15.1.0",
-    "prettier": "^2.8.1",
-    "rimraf": "^3.0.2",
-    "rollup": "^3.7.1",
-    "sinon": "^15.0.0",
+    "prettier": "^2.8.8",
+    "rimraf": "^5.0.0",
+    "rollup": "^3.21.4",
+    "sinon": "^15.0.4",
     "sinon-chai": "^3.7.0",
     "source-map-support": "^0.5.21",
     "standard-version": "^9.5.0",
     "ts-node": "^10.9.1",
-    "tsd": "^0.25.0",
+    "tsd": "^0.28.1",
     "typescript": "^4.9.4",
     "typescript-cached-transpile": "0.0.6",
     "uuid": "^9.0.0",
