@@ -1,13 +1,13 @@
 {
   "name": "dotenv",
-  "version": "16.0.3",
+  "version": "16.1.4",
   "description": "Loads environment variables from .env file",
   "main": "lib/main.js",
   "types": "lib/main.d.ts",
   "exports": {
     ".": {
-      "require": "./lib/main.js",
       "types": "./lib/main.d.ts",
+      "require": "./lib/main.js",
       "default": "./lib/main.js"
     },
     "./config": "./config.js",
@@ -31,6 +31,7 @@
     "type": "git",
     "url": "git://github.com/motdotla/dotenv.git"
   },
+  "funding": "https://github.com/motdotla/dotenv?sponsor=1",
   "keywords": [
     "dotenv",
     "env",
@@ -43,18 +44,21 @@
   "readmeFilename": "README.md",
   "license": "BSD-2-Clause",
   "devDependencies": {
-    "@types/node": "^17.0.9",
+    "@definitelytyped/dtslint": "^0.0.133",
+    "@types/node": "^18.11.3",
     "decache": "^4.6.1",
-    "dtslint": "^3.7.0",
-    "sinon": "^12.0.1",
-    "standard": "^16.0.4",
+    "sinon": "^14.0.1",
+    "standard": "^17.0.0",
     "standard-markdown": "^7.1.0",
-    "standard-version": "^9.3.2",
-    "tap": "^15.1.6",
+    "standard-version": "^9.5.0",
+    "tap": "^16.3.0",
     "tar": "^6.1.11",
-    "typescript": "^4.5.4"
+    "typescript": "^4.8.4"
   },
   "engines": {
     "node": ">=12"
+  },
+  "browser": {
+    "fs": false
   }
 }
