@@ -51,52 +51,50 @@
   "devDependencies": {
     "@iarna/toml": "^2.2.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.33.7",
+    "@microsoft/api-extractor": "^7.34.8",
     "@microsoft/tsdoc-config": "^0.16.2",
-    "@mongodb-js/zstd": "^1.0.0",
-    "@types/chai": "^4.3.3",
+    "@mongodb-js/zstd": "^1.1.0",
+    "@types/chai": "^4.3.5",
     "@types/chai-subset": "^1.3.3",
-    "@types/express": "^4.17.14",
-    "@types/kerberos": "^1.1.1",
-    "@types/mocha": "^10.0.0",
-    "@types/node": "^18.11.0",
+    "@types/express": "^4.17.17",
+    "@types/kerberos": "^1.1.2",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^20.1.0",
     "@types/saslprep": "^1.0.1",
-    "@types/semver": "^7.3.12",
-    "@types/sinon": "^10.0.13",
-    "@types/sinon-chai": "^3.2.8",
+    "@types/semver": "^7.5.0",
+    "@types/sinon": "^10.0.14",
+    "@types/sinon-chai": "^3.2.9",
     "@types/whatwg-url": "^11.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.48.0",
-    "@typescript-eslint/parser": "^5.48.0",
-    "chai": "^4.3.6",
+    "@typescript-eslint/eslint-plugin": "^5.59.5",
+    "@typescript-eslint/parser": "^5.59.5",
+    "chai": "^4.3.7",
     "chai-subset": "^1.6.0",
     "chalk": "^4.1.2",
-    "eslint": "^8.31.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
+    "eslint": "^8.40.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-simple-import-sort": "^8.0.0",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-tsdoc": "^0.2.17",
     "express": "^4.18.2",
     "js-yaml": "^4.1.0",
-    "mocha": "^9.2.2",
+    "mocha": "^10.2.0",
     "mocha-sinon": "^2.1.2",
-    "mongodb-legacy": "^4.0.0",
+    "mongodb-legacy": "^5.0.0",
     "nyc": "^15.1.0",
-    "prettier": "^2.7.1",
-    "rimraf": "^3.0.2",
-    "semver": "^7.3.8",
-    "sinon": "^13.0.1",
+    "prettier": "^2.8.8",
+    "semver": "^7.5.0",
+    "sinon": "^15.0.4",
     "sinon-chai": "^3.7.0",
     "snappy": "^7.2.2",
     "source-map-support": "^0.5.21",
     "standard-version": "^9.5.0",
     "ts-node": "^10.9.1",
-    "tsd": "^0.25.0",
-    "typescript": "^4.9.4",
+    "tsd": "^0.27.0",
+    "typescript": "^4.9.5",
     "typescript-cached-transpile": "^0.0.6",
     "v8-heapsnapshot": "^1.2.0",
-    "xml2js": "^0.4.23",
-    "yargs": "^17.6.0"
+    "yargs": "^17.7.2"
   },
   "license": "Apache-2.0",
   "engines": {
