@@ -39,34 +39,36 @@
     "axios": "1.1.3",
     "babel-loader": "8.2.5",
     "benchmark": "2.1.4",
+    "bluebird": "3.7.2",
     "broken-link-checker": "^0.7.8",
     "buffer": "^5.6.0",
     "cheerio": "1.0.0-rc.12",
     "crypto-browserify": "3.12.0",
-    "dotenv": "16.0.3",
+    "dotenv": "16.1.3",
     "dox": "1.0.0",
-    "eslint": "8.35.0",
+    "eslint": "8.41.0",
     "eslint-plugin-markdown": "^3.0.0",
     "eslint-plugin-mocha-no-only": "1.1.1",
     "express": "^4.18.1",
-    "highlight.js": "11.7.0",
+    "highlight.js": "11.8.0",
     "lodash.isequal": "4.5.0",
     "lodash.isequalwith": "4.4.0",
-    "marked": "4.2.12",
+    "marked": "4.3.0",
     "mkdirp": "^2.1.3",
     "mocha": "10.2.0",
     "moment": "2.x",
-    "mongodb-memory-server": "8.11.5",
+    "mongodb-memory-server": "8.12.2",
     "ncp": "^2.0.0",
     "nyc": "15.1.0",
     "pug": "3.0.2",
     "q": "1.5.1",
-    "sinon": "15.0.1",
+    "sinon": "15.1.0",
     "stream-browserify": "3.0.0",
-    "tsd": "0.25.0",
-    "typescript": "4.9.5",
+    "tsd": "0.28.1",
+    "typescript": "5.1.3",
     "uuid": "9.0.0",
-    "webpack": "5.75.0"
+    "webpack": "5.85.0",
+    "fs-extra": "~11.1.1"
   },
   "directories": {
     "lib": "./lib/mongoose"
@@ -114,7 +118,7 @@
   "main": "./index.js",
   "types": "./types/index.d.ts",
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=14.20.1"
   },
   "bugs": {
     "url": "https://github.com/Automattic/mongoose/issues/new"
