@@ -1,4 +1,5 @@
-const mongoose = require('mongoose')
+
+import mongoose from "mongoose"
 
 const {Schema} = mongoose
 
@@ -33,4 +34,5 @@ const ClassSchema = new Schema({
 })
 
 
-module.exports = mongoose.model('Class', ClassSchema)
+const Class = mongoose.model('Class', ClassSchema)
+export default Class
