@@ -1,14 +1,9 @@
-const express = require("express");
+import express from "express"
 const router = express.Router();
 
-const {
-    getallclasses,
-    createclass,
-    getclass,
-    updateclass,
-    deleteclass,
-} = require("../controller/classes")
+
+import {getallclasses,createclass, getclass, updateclass, deleteclass} from "../controller/classes.js"
 
 router.route('/').get(getallclasses).post(createclass)
 router.route('/:id').get(getclass).patch(updateclass).delete(deleteclass)
-module.exports = router;
+export default router;
